% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelicseries-package.R
\docType{package}
\name{AllelicSeries-package}
\alias{AllelicSeries}
\alias{AllelicSeries-package}
\title{AllelicSeries: Allelic Series Test}
\description{
Implementation of gene-level rare variant association tests targeting allelic series: genes where increasingly deleterious mutations have increasingly large phenotypic effects. The COding-variant Allelic Series Test (COAST) operates on the benign missense variants (BMVs), deleterious missense variants (DMVs), and protein truncating variants (PTVs) within a gene. COAST uses a set of adjustable weights that tailor the test towards rejecting the null hypothesis for genes where the average magnitude of effect increases monotonically from BMVs to DMVs to PTVs. See McCaw ZR, O’Dushlaine C, Somineni H, Bereket M, Klein C, Karaletsos T, Casale FP, Koller D, Soare TW. (2023) "An allelic series rare variant association test for candidate gene discovery" \doi{10.1016/j.ajhg.2023.07.001}.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@insitro.com} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

Other contributors:
\itemize{
  \item Christoph Klein \email{cklein@insitro.com} (\href{https://orcid.org/0000-0002-1748-625X}{ORCID}) [contributor]
  \item Jianhui Gao \email{jianhui.gao@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-0915-1473}{ORCID}) [contributor]
  \item insitro [copyright holder]
}

}
\keyword{internal}
