% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{CalcRegParam}
\alias{CalcRegParam}
\title{Calculate Regression Parameters}
\usage{
CalcRegParam(pve_age = 0.1, pve_pcs = 0.2, pve_sex = 0.1)
}
\arguments{
\item{pve_age}{PVE by age.}

\item{pve_pcs}{PVE by PCs (collectively).}

\item{pve_sex}{PVE by sex.}
}
\value{
List containing the (5 x 1) regression coefficient vector "coef" and
the residual standard deviation "sd".
}
\description{
Calculate phenotypic regression coefficients and the residual variation
based on proportion of variation explained (PVE) by each factor.
}
