% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CatCor}
\alias{CatCor}
\title{Category Correlation}
\usage{
CatCor(anno, ld, maf)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{ld}{(snps x snps) matrix of correlations among the genetic variants.}

\item{maf}{(snps x 1) vector of minor allele frequencies.}
}
\description{
Calculates an estimate of the correlations across the total burdens
of rare variants within each category {BMV, DMV, PTV}. To enable this
calculation, an assumption of Hardy-Weinberg Equilibrium is made, which
provides good performance empirically.
}
