% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{CheckInputs}
\alias{CheckInputs}
\title{Check Inputs}
\usage{
CheckInputs(anno, covar, geno, is_pheno_binary, pheno, weights)
}
\arguments{
\item{anno}{(snps x 1) annotation vector.}

\item{covar}{(n x p) covariate matrix.}

\item{geno}{(n x snps) genotype matrix.}

\item{is_pheno_binary}{Is the phenotype binary?}

\item{pheno}{(n x 1) phenotype vector.}

\item{weights}{(L x 1) annotation category weights.}
}
\value{
None.
}
\description{
Check Inputs
}
