% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{COAST-class}
\alias{COAST-class}
\title{Allelic Series Output Class}
\description{
Allelic Series Output Class
}
\section{Slots}{

\describe{
\item{\code{Betas}}{Effect sizes and standard errors.}

\item{\code{Counts}}{Allele, variant, and carrier counts.}

\item{\code{Pvals}}{Result p-values.}
}}

