% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{show,COAST-method}
\alias{show,COAST-method}
\title{Show Method for COAST Object}
\usage{
\S4method{show}{COAST}(object)
}
\arguments{
\item{object}{An object of class \code{COAST}.}
}
\description{
Show Method for COAST Object
}
