% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ContainsInt}
\alias{ContainsInt}
\title{Contains Intercept?}
\usage{
ContainsInt(x)
}
\arguments{
\item{x}{Design matrix.}
}
\value{
Logical.
}
\description{
Check if a design matrix contains an intercept.
}
