% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CorCpp}
\alias{CorCpp}
\title{Correlation C++}
\usage{
CorCpp(x)
}
\arguments{
\item{x}{Numeric matrix.}
}
\value{
Numeric matrix of correlation among the columns.
}
\description{
Correlation C++
}
\section{Notes}{

Verified this function is faster that R's built-in correlation function
for large genotype matrices.
}

