% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CountAlleles}
\alias{CountAlleles}
\title{Count Alleles}
\usage{
CountAlleles(anno, geno, count_carriers = FALSE, min_mac = 0)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{geno}{(n x snps) genotype matrix.}

\item{count_carriers}{If true, counts the number of carriers rather than
the number of alleles.}

\item{min_mac}{Minimum minor allele count for inclusion. Default: 0.}
}
\value{
3 x 1 numeric vector with the counts of BMVs, DMVs, and PTVs.
}
\description{
Count the number of non-zero alleles bearing each variant annotation.
}
