% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{FilterGenos}
\alias{FilterGenos}
\title{Filter Noncausal Variants}
\usage{
FilterGenos(anno, geno, prop_causal = 1)
}
\arguments{
\item{anno}{(snps x 1) annotation vector.}

\item{geno}{(n x snps) genotype matrix.}

\item{prop_causal}{Proportion of variants which are causal.}
}
\value{
List containing the (n x snps) genotype matrix "geno" and the
(snps x 1) annotation vector "anno".
}
\description{
Remove a random fraction of variants, which are designated non-causal.
}
