% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenAnno}
\alias{GenAnno}
\title{Generate Genotype Annotations}
\usage{
GenAnno(mat, p_dmv = 0.33, p_ptv = 0.33)
}
\arguments{
\item{mat}{Genotype matrix.}

\item{p_dmv}{Frequency of deleterious missense variants.}

\item{p_ptv}{Frequency of protein truncating variants.}
}
\value{
(snps x 1) integer vector.
}
\description{
Returns a vector of length = the number of columns (SNPs) in the genotype
matrix. Each SNP is classified as a benign missense variant (0), a
deleterious missense variant (1), or a protein truncating variant (2).
}
