% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenCovar}
\alias{GenCovar}
\title{Generate Covariates}
\usage{
GenCovar(n)
}
\arguments{
\item{n}{Sample size.}
}
\value{
(n x 6) numeric matrix.
}
\description{
Generate an (n x 6) covariate matrix with columns representing an intercept,
age, sex, and 3 genetic PCs. Because these simulations address rare variant
analysis, correlation between genotypes and the genetic PCs (based on
common variants) is unnecessary.
}
