% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenGeno}
\alias{GenGeno}
\title{Generate Genotypes}
\usage{
GenGeno(n, snps, maf_range = c(0.005, 0.01), p_dmv = 0.33, p_ptv = 0.33)
}
\arguments{
\item{n}{Sample size.}

\item{snps}{Number of SNP in the gene.}

\item{maf_range}{Range of minor allele frequencies: c(MIN, MAX).}

\item{p_dmv}{Frequency of deleterious missense variants.}

\item{p_ptv}{Frequency of protein truncating variants.}
}
\value{
List containing the (n x snps) genotype matrix "geno" and the
(snps x 1) annotation vector "anno".
}
\description{
Generate Genotypes
}
