% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenomicControl}
\alias{GenomicControl}
\title{Genomic Control}
\usage{
GenomicControl(lambda, pval, df = 1)
}
\arguments{
\item{lambda}{Genomic inflation factor.}

\item{pval}{Numeric p-value.}

\item{df}{Degrees of freedom. Should not require modification in most cases.}
}
\value{
Corrected p-value.
}
\description{
Genomic Control
}
