% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isPD}
\alias{isPD}
\title{Check if Positive Semi Definite}
\usage{
isPD(x, force_symmetry = FALSE)
}
\arguments{
\item{x}{Numeric matrix.}

\item{force_symmetry}{Force the matrix to be symmetric?}
}
\value{
Logical indicating whether the matrix is SPD.
}
\description{
Check if Positive Semi Definite
}
