% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IVWCpp}
\alias{IVWCpp}
\title{Inverse Variance Meta-Analysis}
\usage{
IVWCpp(anno, beta, se, ld, weights)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{beta}{(snps x 1) vector of effect sizes for
the coding genetic variants within a gene.}

\item{se}{(snps x 1) vector of standard errors for the effect sizes.}

\item{ld}{(snps x snps) matrix of correlations among the genetic variants.}

\item{weights}{(3 x 1) vector of annotation category weights.}
}
\value{
Data.frame with the following columns \itemize{
\item anno: gene annotation
\item beta_meta: meta-analyzed effect size
\item se_meta: standard error of the meta-analyzed effect size
}
}
\description{
Inverse Variance Meta-Analysis
}
