% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{IVWSS}
\alias{IVWSS}
\title{Inverse Variance Meta-Analysis}
\usage{
IVWSS(anno, beta, se, ld = NULL, weights = c(1, 2, 3))
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{beta}{(snps x 1) vector of effect sizes for
the coding genetic variants within a gene.}

\item{se}{(snps x 1) vector of standard errors for the effect sizes.}

\item{ld}{(snps x snps) matrix of correlations among the genetic variants.
Defaults to the identity matrix.}

\item{weights}{Annotation category weights. Defaults to c(1, 2, 3).}
}
\value{
Data.frame with the following columns \itemize{
\item anno: gene annotation
\item beta_meta: meta-analyzed effect size
\item se_meta: standard error of the meta-analyzed effect size
}
}
\description{
Inverse Variance Meta-Analysis
}
\section{Notes}{

This wrapper function is needed because C++ does not allow for default
arguments with dynamically allocated values.
}

