% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.COAST}
\alias{print.COAST}
\title{Print Method for COAST Object.}
\usage{
\method{print}{COAST}(x, ...)
}
\arguments{
\item{x}{An object of class \code{COAST}.}

\item{...}{Unused.}
}
\description{
Print method for objects of class \code{COAST}.
}
