% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ResidVar}
\alias{ResidVar}
\title{Calculate Residual Variance}
\usage{
ResidVar(y, X)
}
\arguments{
\item{y}{(n x 1) Numeric phenotype vector.}

\item{X}{(n x q) Numeric covariate matrix.}
}
\value{
Scalar residual variance.
}
\description{
Calculate Residual Variance
}
