% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Score}
\alias{Score}
\title{Calculate Score Statistic}
\usage{
Score(y, G, X, v)
}
\arguments{
\item{y}{(n x 1) Numeric phenotype vector.}

\item{G}{(n x p) Numeric genotype matrix.}

\item{X}{(n x q) Numeric covariate matrix.}

\item{v}{Scalar residual variance.}
}
\value{
Scalar score statistic.
}
\description{
Calculate Score Statistic
}
