% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_series_test.R
\name{Aggregator}
\alias{Aggregator}
\title{Aggregator}
\usage{
Aggregator(
  anno,
  geno,
  drop_empty = TRUE,
  indicator = FALSE,
  method = "none",
  weights = DEFAULT_WEIGHTS
)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{geno}{(n x snps) genotype matrix.}

\item{drop_empty}{Drop empty columns? Default: TRUE.}

\item{indicator}{Convert raw counts to indicators? Default: FALSE.}

\item{method}{Method for aggregating across categories:
{"none", "max", "sum"}. Default: "none".}

\item{weights}{Annotation category weights.}
}
\value{
(n x 3) Numeric matrix without weighting, (n x 1) numeric matrix
with weighting.
}
\description{
Aggregates genotypes within annotation categories.
}
