% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allelic_series_test.R
\name{COAST}
\alias{COAST}
\title{COding-variant Allelic Series Test}
\usage{
COAST(
  anno,
  geno,
  pheno,
  apply_int = TRUE,
  covar = NULL,
  include_orig_skato_all = FALSE,
  include_orig_skato_ptv = FALSE,
  is_pheno_binary = FALSE,
  return_omni_only = FALSE,
  weights = DEFAULT_WEIGHTS
)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with values in c(0, 1, 2).}

\item{geno}{(n x snps) genotype matrix.}

\item{pheno}{(n x 1) phenotype vector.}

\item{apply_int}{Apply rank-based inverse normal transform to the phenotype?
Default: TRUE. Ignored if phenotype is binary.}

\item{covar}{(n x p) covariate matrix. Defaults to an (n x 1) intercept.}

\item{include_orig_skato_all}{Include the original version of SKAT-O applied
to all variants in the omnibus test? Default: FALSE.}

\item{include_orig_skato_ptv}{Include the original version of SKAT-O applied
to PTV variants only in the omnibus test? Default: FALSE.}

\item{is_pheno_binary}{Is the phenotype binary? Default: FALSE.}

\item{return_omni_only}{Return only the omnibus p-value? Default: FALSE.}

\item{weights}{(3 x 1) annotation category weights.}
}
\value{
Numeric p-value.
}
\description{
Main allelic series test. Performs both Burden and SKAT type tests, then
combines the results to calculate an omnibus p-value.
}
\examples{
# Generate data.
data <- DGP(n = 1e3, snps = 1e2)

# Run the COding-variant Allelic Series Test.
results <- COAST(
  anno = data$anno,
  geno = data$geno,
  pheno = data$pheno,
  covar = data$covar
)
show(results)

}
