% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Counts}
\alias{Counts}
\title{Count Variants and Carriers}
\usage{
Counts(anno, geno, n_anno, min_mac = 0L)
}
\arguments{
\item{anno}{(snps x 1) annotation vector with integer values in 1 through
the number of annotation categories L.}

\item{geno}{(n x snps) genotype matrix.}

\item{n_anno}{Number of annotation categories L.}

\item{min_mac}{Minimum minor allele count for inclusion. Default: 0.}
}
\value{
Data.frame of allele, variant, and carrier counts.
}
\description{
Count Variants and Carriers
}
