% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenGenoMat}
\alias{GenGenoMat}
\title{Generate Genotype Matrix}
\usage{
GenGenoMat(n, snps, maf_range = c(0.001, 0.005))
}
\arguments{
\item{n}{Sample size.}

\item{snps}{Number of SNP in the gene.}

\item{maf_range}{Range of minor allele frequencies: c(MIN, MAX).}
}
\value{
(n x snps) numeric matrix.
}
\description{
Generate genotypes for \code{n} subject at \code{snps} variants in linkage
equilibrium. Genotypes are generated such that the MAC is always >= 1.
}
