% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{GenPheno}
\alias{GenPheno}
\title{Generate Phenotypes}
\usage{
GenPheno(
  anno,
  beta,
  covar,
  geno,
  reg_param,
  binary = FALSE,
  include_residual = TRUE,
  indicator = FALSE,
  method = "none",
  random_signs = FALSE,
  weights = c(0, 1, 2)
)
}
\arguments{
\item{anno}{(snps x 1) annotation vector.}

\item{beta}{(3 x 1) coefficient vector for bmvs, dmvs, and ptvs respectively.}

\item{covar}{Covariate matrix.}

\item{geno}{(n x snps) genotype matrix.}

\item{reg_param}{Regression parameters.}

\item{binary}{Generate binary phenotype? Default: FALSE.}

\item{include_residual}{Include residual? If FALSE, returns the expected
value. Intended for testing.}

\item{indicator}{Convert raw counts to indicators? Default: FALSE.}

\item{method}{Genotype aggregation method. Default: "none".}

\item{random_signs}{Randomize signs? FALSE for burden-type genetic
architecture, TRUE for SKAT-type.}

\item{weights}{Aggregation weights.}
}
\value{
(n x 1) numeric vector.
}
\description{
Generate Phenotypes
}
