% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLS}
\alias{OLS}
\title{Ordinary Least Squares}
\usage{
OLS(y, X)
}
\arguments{
\item{y}{(n x 1) Numeric vector.}

\item{X}{(n x p) Numeric matrix.}
}
\value{
List containing the following:
\itemize{
\item{beta: Regression coefficients.}
\item{v: Residual variance.}
\item{se: Standard errors.}
\item{z: Z-scores.}
\item{pval: P-values based on the chi2 distribution.}
}
}
\description{
Fits the standard OLS model.
}
