% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SumCountSS}
\alias{SumCountSS}
\title{Allelic Sum Test from Sumstats}
\usage{
SumCountSS(anno, beta, ld, se, weights)
}
\arguments{
\item{anno}{(snps x 1) annotation vector.}

\item{beta}{(snps x 1) vector of effect sizes for
the coding genetic variants within a gene.}

\item{ld}{(snps x snps) matrix of correlations among the genetic variants.}

\item{se}{(snps x 1) vector of standard errors for the effect sizes.}

\item{weights}{(3 x 1) vector of annotation category weights.}
}
\value{
Numeric p-value.
}
\description{
Allelic Sum Test from Sumstats
}
